Add oath2 file, update linter, construct Implementations map.
