+++ lib/main.js
@@ -1,4 +1,10 @@
 /*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2022 Digital Bazaar, Inc. All rights reserved.
  */
-export {Example} from './Example.js';
+import * as implementorFiles from '../implementations';
+import Implementation from './Implementation';
+
+const keyValues = implementorFiles.map(implementation => [implementation.name, new Implementation(implementation)]);
+export const implementations = new Map(keyValues);
+export default implementations;
+
